+++ gdbgui/src/js/gdbgui.jsx
@@ -10,20 +10,22 @@
  /* global debug */
  /* global initial_data */
 
-import {store, initial_store_data} from './store.js';
-import GdbApi from './GdbApi.jsx';
-import ReactDOM from 'react-dom';
-import React from 'react';
-import TopBar from './TopBar.jsx';
-import GlobalEvents from './GlobalEvents.js';
-import MiddleLeft from './MiddleLeft.jsx';
-import FileOps from './FileOps.jsx';
-import Settings from './Settings.jsx';
-import Modal from './GdbguiModal.jsx';
-import HoverVar from './HoverVar.jsx';
-import RightSidebar from './RightSidebar.jsx';
-import FoldersView from './FoldersView.jsx';
-import GdbConsoleContainer from './GdbConsoleContainer.jsx';
+import {store, initial_store_data} from './store.js'
+import GdbApi from './GdbApi.jsx'
+import ReactDOM from 'react-dom'
+import React from 'react'
+import TopBar from './TopBar.jsx'
+import GlobalEvents from './GlobalEvents.js'
+import MiddleLeft from './MiddleLeft.jsx'
+import FileOps from './FileOps.jsx'
+import Settings from './Settings.jsx'
+import Modal from './GdbguiModal.jsx'
+import HoverVar from './HoverVar.jsx'
+import ToolTip from './ToolTip.jsx'
+import RightSidebar from './RightSidebar.jsx'
+import FoldersView from './FoldersView.jsx'
+import GdbConsoleContainer from './GdbConsoleContainer.jsx'
+import Actions from './Actions.js'
 
 store.options.debug = debug
 store.initialize(initial_store_data)
@@ -70,9 +72,15 @@ class Gdbgui extends React.PureComponent {
                   </div>
                 </div>
 
+                {/* below are elements that are only displayed under certain conditions */}
                 <Modal />
                 <HoverVar />
                 <Settings />
+                <ToolTip />
+                <textarea ref={(node)=>{
+                    store.set('textarea_to_copy_to_clipboard', node)
+                  }
+                } />
             </div>
         )
     }
@@ -95,6 +103,28 @@ class Gdbgui extends React.PureComponent {
         })
 
         store.set('middle_panes_split_obj', middle_panes_split_obj)
+
+        // Fetch the latest version only if using in normal mode. If debugging, we tend to
+        // refresh quite a bit, which might make too many requests to github and cause them
+        // to block our ip? Either way it just seems weird to make so many ajax requests.
+        if(!store.get('debug')){
+            // fetch version
+            $.ajax({
+                url: "https://raw.githubusercontent.com/cs01/gdbgui/master/gdbgui/VERSION.txt",
+                cache: false,
+                method: 'GET',
+                success: (data) => {
+                    store.set('latest_gdbgui_version', _.trim(data))
+                    if(TopBar.needs_to_update_gdbgui_version() && store.get('show_gdbgui_upgrades')){
+                        Actions.show_modal(`Update Available`, Settings.get_upgrade_text())
+                    }
+                },
+                error: (data) => {
+                    void(data)
+                    store.set('latest_gdbgui_version', '(could not contact server)')
+                }
+            })
+        }
     }
 
 }