+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Fix safari ui issue (#164)
 * Update text on reload file button, and disable when no file is loaded (#165)
 * When disassembly can't be fetched in mode 4, fetch in mode 3 and assume gdb version is 7.6.0 (#166)
+* Add copy to clipboard icon for files
+* Allow SSL module import to fail and print warning (#170)
 * Cleanup menu, add license info, etc.
 
 ## 0.11.0.0